# Faster (almost 2x) mutexes (was slower due to ThreadFuzzer)