### Changelog category (leave one):
- Performance Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Faster (almost 2x) mutexes (was slower due to ThreadFuzzer)

Recently I noticed that DB::injection() pops up in perf top, so let's optimize it slightly:
- Add -fomit-frame-pointer -momit-leaf-frame-pointer explicitly -- almost 0 effect
- Add ALWAYS_INLINE for ThreadFuzzer::isStarted() (just in case)
- Disable ThreadFuzzer if non of env variables had been set, this is needed to avoid extra checks in DB::injection()
- Add ALWAYS_INLINE for ThreadFuzzer::injection()

And here are some results for ThreadFuzzer test:
- before:   elapsed 6.27368 / test time 654 ms
- after:    elapsed 3.14167 / test time 325 ms
- disabled: elapsed 2.46666 / test time 263 ms

*But note, it is still slower then with ThreadFuzzer disabled.*

Note, that this numbers for AMD 5975WX, for server with 2x Xeon Silver 4216 2.10:
- before:   elapsed 12.109  / test time 1325 ms
- after:    elapsed 10.506  / test time 1101 ms
- disabled: elapsed 8.41043 / test time 917 ms

P.S. I've also tried with the same glibc version as server had - zero changes.

Refs: https://gist.github.com/azat/51a5fcc3a40af9f678906a3a6e14e079