@azat 2024-03-05
There are few very common loggers (per-table), initially I looked into this while looking at https://github.com/ClickHouse/ClickHouse/pull/59531

And with this patch I got:
>localhost:9000, queries: 6416, QPS: 413.256, RPS: 413.256, MiB/s: 0.006, result RPS: 0.000, result MiB/s: 0.000.

While before it was:
>localhost:9000, queries: 4463, QPS: 375.140, RPS: 375.140, MiB/s: 0.006, result RPS: 0.000, result MiB/s: 0.000.

With empty `ReplicatedMergeTreeQueue::getAlterMutationCommandsForPart`:
>localhost:9000, queries: 8588, QPS: 653.918, RPS: 653.918, MiB/s: 0.010, result RPS: 0.000, result MiB/s: 0.000.

With ThreadFuzzer completely compiled out:
>localhost:9000, queries: 7745, QPS: 434.357, RPS: 434.357, MiB/s: 0.007, result RPS: 0.000, result MiB/s: 0.000.

Without mutations:
>localhost:9000, queries: 7381, QPS: 520.910, RPS: 520.910, MiB/s: 0.008, result RPS: 0.000, result MiB/s: 0.000.