@azat 2024-03-05
>a little silly question: it shouldn't be enabled on servers running prod workloads, so what are we optimising?

It is - https://github.com/ClickHouse/ClickHouse/blob/af3b16a79d4d7397d9dc2da5b74f4b57a8de8c39/src/Common/ThreadFuzzer.cpp#L26

Since the same binary used for CI and for releases