@alexey-milovidov 2024-03-05
Yes, we should run the same binary in production and in CI.
It is good to be able to use ThreadFuzzer in production builds.